Refactor: use other service in ReviewController
