+++ src/main/java/com/devthink/devthink_server/application/ReviewService.java
@@ -36,10 +36,7 @@ public class ReviewService {
      * @return 생성 된 리뷰 id
      */
     @Transactional
-    public String createReview(ReviewRequestData reviewRequestData) {
-        User user = userRepository.findByIdAndDeletedIsFalse(reviewRequestData.getUserId())
-                .orElseThrow(() -> new UserNotFoundException(reviewRequestData.getUserId()));
-        Book book = getBookByIsbn(reviewRequestData.getBook());
+    public String createReview(User user, Book book, ReviewRequestData reviewRequestData) {
 
         if (!reviewRepository.findByBookAndUser(book, user).isEmpty()) { // 이미 작성한 리뷰인지 확인합니다.
             throw new AlreadyReviewedException(book.getId());
@@ -57,39 +54,6 @@ public class ReviewService {
         review.getBook().setScoreAvg(bookRepository.calcScoreAvg(book.getId()));
         return review.getId().toString();
     }
-
-    /**
-     * 입력된 isbn 정보로 Book을 조회하며, 해당 책이 없는 경우 새로 생성하는 함수를 호출합니다.
-     *
-     * @param bookRequestData (책에 대한 정보)
-     * @return 조회 혹은 생성된 Book 객체
-     */
-    public Book getBookByIsbn(BookRequestData bookRequestData) {
-        Optional<Book> book = bookRepository.getBookByIsbn(bookRequestData.getIsbn());
-        if (book.isEmpty()) {
-            return createBook(bookRequestData);
-        } else {
-            return book.get();
-        }
-    }
-
-    /**
-     * 입력된 책 정보의 새로운 Book 객체를 생성합니다.
-     *
-     * @param bookRequestData (책에 대한 정보)
-     * @return 생성된 Book 객체
-     */
-    @Transactional
-    public Book createBook(BookRequestData bookRequestData) {
-        Book book = Book.builder()
-                .isbn(bookRequestData.getIsbn())
-                .name(bookRequestData.getName())
-                .writer(bookRequestData.getWriter())
-                .imgUrl(bookRequestData.getImgUrl())
-                .build();
-        return bookRepository.save(book);
-    }
-
 
     /**
      * 입력된 리뷰 식별자 (id) 값으로 리뷰 상세 정보를 가져옵니다.
